🚚 Move `request_to_investing` to utils
