+++ src/investiny/utils.py
@@ -1,0 +1,32 @@
+# Copyright 2022 Alvaro Bartolome, alvarobartt @ GitHub
+# See LICENSE for details.
+
+import json
+from typing import Any, Dict
+from uuid import uuid4
+
+import httpx
+
+__all__ = ["request_to_investing"]
+
+
+def request_to_investing(params: Dict[str, Any]) -> Dict[str, Any]:
+    """Sends an HTTP GET request to Investing.com API with the introduced params.
+
+    Args:
+        params (Dict[str, Any]): A dictionary with the params to send to Investing.com API.
+
+    Returns:
+        Dict[str, Any]: A dictionary with the response from Investing.com API.
+    """
+    url = f"https://tvc4.investing.com/{uuid4().hex}/0/0/0/0/history"
+    headers = {
+        "User-Agent": (
+            "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like"
+            " Gecko) Chrome/104.0.5112.102 Safari/537.36"
+        ),
+        "Referer": "https://tvc-invdn-com.investing.com/",
+        "Content-Type": "application/json",
+    }
+    r = httpx.get(url, params=params, headers=headers)
+    return json.loads(r.text)  # type: ignore
